Conflicts:
	HelpMe.md
	diceparser.cpp
	parsingtoolbox.cpp
	result/result.cpp
	result/stringresult.h

+++ validator.h
@@ -54,7 +54,7 @@ public:
      * @param faces
      * @return
      */
-    virtual quint8 getValidRangeSize(quint64 faces) const = 0 ;
+    virtual quint64 getValidRangeSize(quint64 faces) const = 0 ;
 };
 
 #endif // VALIDATOR_H
